[SDSAET-38] Move CookieError models to models-api module
