+++ api/models-api/src/main/java/com/cognifide/aet/models/accessibility/AccessibilityIssue.java
@@ -13,8 +13,9 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.models;
+package com.cognifide.aet.models.accessibility;
 
+import com.cognifide.aet.models.Excludable;
 import java.io.Serializable;
 import java.util.Objects;
