+++ api/models-api/src/main/java/com/cognifide/aet/models/accessibility/AccessibilityReport.java
@@ -13,7 +13,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.models;
+package com.cognifide.aet.models.accessibility;
 
 import java.io.Serializable;
 import java.util.List;
