+++ api/models-api/src/main/java/com/cognifide/aet/models/accessibility/AccessibilityReportConfiguration.java
@@ -13,9 +13,9 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.models;
+package com.cognifide.aet.models.accessibility;
 
-import com.cognifide.aet.models.AccessibilityIssue.IssueType;
+import com.cognifide.aet.models.accessibility.AccessibilityIssue.IssueType;
 import java.util.Map;
 import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
