+++ core/jobs/src/main/java/com/cognifide/aet/job/common/comparators/accessibility/AccessibilityComparator.java
@@ -22,10 +22,10 @@ import com.cognifide.aet.job.api.datafilter.DataFilterJob;
 import com.cognifide.aet.job.api.exceptions.ParametersException;
 import com.cognifide.aet.job.api.exceptions.ProcessingException;
 import com.cognifide.aet.job.common.comparators.accessibility.report.AccessibilityReportGenerator;
-import com.cognifide.aet.models.AccessibilityIssue;
-import com.cognifide.aet.models.AccessibilityReport;
-import com.cognifide.aet.models.AccessibilityReportConfiguration;
 import com.cognifide.aet.models.Excludable;
+import com.cognifide.aet.models.accessibility.AccessibilityIssue;
+import com.cognifide.aet.models.accessibility.AccessibilityReport;
+import com.cognifide.aet.models.accessibility.AccessibilityReportConfiguration;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.Lists;
