+++ core/jobs/src/main/java/com/cognifide/aet/job/common/comparators/accessibility/report/AccessibilityReportGenerator.java
@@ -15,10 +15,10 @@
  */
 package com.cognifide.aet.job.common.comparators.accessibility.report;
 
-import com.cognifide.aet.models.AccessibilityIssue;
-import com.cognifide.aet.models.AccessibilityIssue.IssueType;
-import com.cognifide.aet.models.AccessibilityReport;
-import com.cognifide.aet.models.AccessibilityReportConfiguration;
+import com.cognifide.aet.models.accessibility.AccessibilityIssue;
+import com.cognifide.aet.models.accessibility.AccessibilityIssue.IssueType;
+import com.cognifide.aet.models.accessibility.AccessibilityReport;
+import com.cognifide.aet.models.accessibility.AccessibilityReportConfiguration;
 import com.google.common.base.Predicate;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
