+++ core/jobs/src/main/java/com/cognifide/aet/job/common/comparators/cookie/CookieComparator.java
@@ -20,6 +20,10 @@ import com.cognifide.aet.job.api.comparator.ComparatorJob;
 import com.cognifide.aet.job.api.comparator.ComparatorProperties;
 import com.cognifide.aet.job.api.exceptions.ParametersException;
 import com.cognifide.aet.job.api.exceptions.ProcessingException;
+import com.cognifide.aet.models.cookie.CompareAction;
+import com.cognifide.aet.models.cookie.CookieComparatorResult;
+import com.cognifide.aet.models.cookie.CookieCompareComparatorResult;
+import com.cognifide.aet.models.cookie.CookieTestComparatorResult;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import com.google.common.base.Function;
 import com.google.common.collect.FluentIterable;
