+++ core/jobs/src/test/java/com/cognifide/aet/job/common/collectors/accessibility/AccessibilityIssueMarkupFinderTest.java
@@ -17,7 +17,7 @@ package com.cognifide.aet.job.common.collectors.accessibility;
 
 import static org.junit.Assert.assertEquals;
 
-import com.cognifide.aet.models.AccessibilityIssue;
+import com.cognifide.aet.models.accessibility.AccessibilityIssue;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.Before;
