+++ rest-endpoint/src/main/java/com/cognifide/aet/repositories/ErrorWrapperRepository.java
@@ -17,15 +17,9 @@ package com.cognifide.aet.repositories;
 
 import com.cognifide.aet.communication.api.metadata.Comparator;
 import com.cognifide.aet.job.api.collector.JsErrorLog;
-import com.cognifide.aet.job.common.comparators.cookie.CookieCompareComparatorResult;
-import com.cognifide.aet.job.common.comparators.cookie.CookieTestComparatorResult;
 import com.cognifide.aet.job.common.comparators.source.diff.ResultDelta;
 import com.cognifide.aet.job.common.comparators.statuscodes.StatusCodesComparatorResult;
 import com.cognifide.aet.job.common.comparators.w3chtml5.W3cHtml5ComparatorResult;
-import com.cognifide.aet.models.AccessibilityErrorWrapper;
-import com.cognifide.aet.models.AccessibilityReport;
-import com.cognifide.aet.models.AccessibilityReportConfiguration;
-import com.cognifide.aet.models.CookieErrorWrapper;
 import com.cognifide.aet.models.ErrorType;
 import com.cognifide.aet.models.ErrorWrapper;
 import com.cognifide.aet.models.JsErrorWrapper;
@@ -33,6 +27,12 @@ import com.cognifide.aet.models.ScreenErrorWrapper;
 import com.cognifide.aet.models.SourceErrorWrapper;
 import com.cognifide.aet.models.StatusCodesErrorWrapper;
 import com.cognifide.aet.models.W3cHtml5ErrorWrapper;
+import com.cognifide.aet.models.accessibility.AccessibilityErrorWrapper;
+import com.cognifide.aet.models.accessibility.AccessibilityReport;
+import com.cognifide.aet.models.accessibility.AccessibilityReportConfiguration;
+import com.cognifide.aet.models.cookie.CookieCompareComparatorResult;
+import com.cognifide.aet.models.cookie.CookieErrorWrapper;
+import com.cognifide.aet.models.cookie.CookieTestComparatorResult;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import com.cognifide.aet.vs.DBKey;
 import java.io.IOException;
